+++ src/mkvtoolnix-gui/util/settings_names.h
@@ -90,6 +90,7 @@ char const * const s_valMergePredefinedSubtitleTrackNames         = "mergePredef
 char const * const s_valMergePredefinedTrackNames                 = "mergePredefinedTrackNames";
 char const * const s_valMergePredefinedVideoTrackNames            = "mergePredefinedVideoTrackNames";
 char const * const s_valMergeReconstructSequencesWhenAdding       = "mergeReconstructSequencesWhenAdding";
+char const * const s_valMergeSetDestinationFromDirectory          = "mergeSetDestinationFromDirectory";
 char const * const s_valMergeSetDestinationFromTitle              = "mergeSetDestinationFromTitle";
 char const * const s_valMergeSortFilesTracksByTypeWhenAdding      = "mergeSortFilesTracksByTypeWhenAdding";
 char const * const s_valMergeTrackPropertiesLayout                = "mergeTrackPropertiesLayout";